+++ packages/async/src/retry.ts
@@ -1,3 +1,5 @@
+import { hoc } from '@starkit/utils/hoc'
+
 import { sleep } from './sleep'
 import { Runnable } from './types'
 
@@ -6,21 +8,21 @@ export function retry<Fn extends Runnable>(
   opts: { times?: number; interval?: number | Runnable } = {}
 ) {
   const { times = 3, interval = 0 } = opts
-  const name = fn.name || 'anonymous'
 
-  return {
-    [name]: async function (this: unknown, s?: AbortSignal) {
+  return hoc(
+    async function (this: unknown, s?: AbortSignal) {
       let i = 0
       while (i++ < times) {
         try {
           return fn(s)
         } catch (e) {
-          if (i === times) throw e
+          if (i >= times) throw e
           await (typeof interval === 'number' ? sleep(interval) : interval(s))
         }
       }
     } as ReturnType<Fn> extends Promise<unknown>
       ? Fn
-      : (...args: Parameters<Fn>) => Promise<ReturnType<Fn>>
-  }[name]
+      : (...args: Parameters<Fn>) => Promise<ReturnType<Fn>>,
+    fn
+  )
 }