+++ packages/async/src/safe.ts
@@ -1,11 +1,11 @@
+import { hoc } from '@starkit/utils/hoc'
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Result<T> = { ok: true; value: T } | { ok: false; reason: any }
 
 export function safe<A extends unknown[], T>(fn: (...args: A) => T) {
-  const name = fn.name || 'anonymous'
-
-  return {
-    [name]: function (this: unknown, ...args: A) {
+  return hoc(
+    function (this: unknown, ...args: A) {
       try {
         const r = fn.apply(this, args)
         if (r instanceof Promise) {
@@ -21,6 +21,7 @@ export function safe<A extends unknown[], T>(fn: (...args: A) => T) {
       }
     } as (
       ...args: A
-    ) => T extends Promise<unknown> ? Promise<Result<Awaited<T>>> : Result<T>
-  }[name]
+    ) => T extends Promise<unknown> ? Promise<Result<Awaited<T>>> : Result<T>,
+    fn
+  )
 }