+++ packages/utils/src/debounce.ts
@@ -1,3 +1,5 @@
+import { hoc } from './hoc'
+
 type Fn<A extends unknown[], R> = (...args: A) => R
 
 export function debounce<A extends unknown[], R>(
@@ -6,13 +8,10 @@ export function debounce<A extends unknown[], R>(
 ): Fn<A, () => void> {
   let t: ReturnType<typeof setTimeout> | undefined
   const dispose = () => void (t && clearTimeout(t))
-  const name = func.name || 'anonymous'
 
-  return {
-    [name]: function (this: unknown, ...args: A) {
-      t && clearTimeout(t)
-      t = setTimeout(func.bind(this), wait, ...args)
-      return dispose
-    }
-  }[name]
+  return hoc(function (this: unknown, ...args: A) {
+    t && clearTimeout(t)
+    t = setTimeout(func.bind(this), wait, ...args)
+    return dispose
+  }, func)
 }